+++ client/src/components/shared/Catbar.js
@@ -1,23 +1,16 @@
 import React, { useState, useEffect } from 'react'
-import { AuthConsumer, } from "../../providers/AuthProvider";
 import { Menu, } from 'semantic-ui-react'
 import { Link, withRouter, } from 'react-router-dom'
-import ShowModal from '../modal/ShowModal'
 import axios from 'axios';
 import styled from 'styled-components'
 
 const Catbar = () => {
-  
   const [cats, setCats] = useState([])
+
   useEffect(() => {
     axios.get("/api/categories")
-      .then( res => {
-        console.log(res.data)
-        setCats(res.data)
-      })
-      .catch( err => {
-        console.log(err)
-      })
+      .then( res => setCats(res.data))
+      .catch(console.log)
   }, [])
   
   return (
@@ -28,7 +21,7 @@ const Catbar = () => {
             <Menu.Item
               name={cat.title}
               id={cat.title}
-              // active={this.props.location.pathname === '/'}
+              // active={this.props.location.pathname === '/'} //START WORKING HERE
             />
           </Link>
       ))}
